pyproject.toml
[project]
name = "matrika"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.23.0",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "email-validator>=2.3.0",
    "google>=3.0.0",
    "google-genai>=1.56.0",
    "jinja2>=3.1.6",
    "openai>=2.11.0",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "sqlalchemy[asyncio]>=2.0.44",
    "yookassa>=3.9.0",
]
